Added phone check
