+++ HW1/README.md
@@ -15,7 +15,7 @@
 
 ### Ход работы:
 
-1. Были созданы файлы function.py и main.py, в которых находится логика приложения
+1. Были созданы файлы function.py, menu.py и main.py, в которых находится логика приложения
 2. Приложение ведет работу с файлом phone_book.json. В случае, если данного файла не существует, то файл создается автоматически
 3. Реализованы методы для работы с телефонным справочником для:
    * добавление контакта
